MD-d74 добавить моковый файл
