+++ public/components/actor/actor.pug
@@ -2,7 +2,7 @@ include ../filmCard/filmCard
 include ../slider/slider
 div.actor-info
     div.actor-info__container
-        img(src=picture_url)
+        img(src=actor.avatar)
         div.actor-info__container_text
             div.actor-info__container_text_name
                 div.actor-info__container_text_name_rus #{actor.name}
