+++ public/components/actor/actor.scss
@@ -28,13 +28,11 @@
         &_rus {
           font-family: "Montserrat-Black", "Arial", sans-serif;
           font-size: 40px;
-          line-height: normal;
           margin-bottom: 12px;
         }
 
         &_en {
           font-size: 16px;
-          line-height: normal;
         }
       }
