+++ public/components/filmCard/filmCard.pug
@@ -1,10 +1,9 @@
 mixin filmCard (i)
     div.film-card
         div.film-card_pic
-            img(src=i.film_avatar)
+            img(src=i.filmAvatar)
         div.film-card_description
             div.film-card_description_name #{i.title}
                 div.film-card_description_name_year #{i.year}
             div.film-card_description_text #{i.description}
             button(type='button' class="pink-button" ) Подробнее
-
