+++ public/views/HomeView/HomeView.js
@@ -1,6 +1,5 @@
 import {ROOT} from '../../main.js';
 import {BaseView} from '../BaseView/BaseView.js';
-import header from '../../components/header/header.pug';
 import homeContent from '../../components/collections/collections.pug';
 import errorPage from '../../components/errorPage/errorPage.pug';
 import {Events} from '../../consts/events.js';
